renamed games to models
